# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW SyncIt! Web API'
  description: 'Web Resources Specification (A7) for SyncIt!'
servers:
- url: http://lbaw24115.lbaw.fe.up.pt
  description: Production server
...
paths:
...
  /tickets:
    get:
      operationId: R204
      summary: 'R204: View user tickets'
      description: 'Show a list of tickets purchased by the user. Access: OWN'
      tags:
        - 'M03: Events and Tickets'
      responses:
        '200':
          description: 'List of tickets retrieved successfully.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      operationId: R205
      summary: 'R205: Refund Ticket'
      description: 'Allows a member to refund a ticket. Access: OWN.'
      tags:
        - 'M03: Events and Tickets'
      requestBody:
        required: true
        description: 'Ticket details for the refund request.'
        content:
          application/json:
            schema:
              type: object
              properties:
                ticket_id:
                  type: integer
                  description: 'Unique identifier of the ticket being refunded.'
                  required: true
      responses:
        '200':
          description: 'Ticket refunded successfully.'
          headers:
            Location:
              description: 'Redirects to the tickets list or the refund success page.'
              schema:
                type: string
                example: '/tickets'

        '400':
          $ref: '#/components/responses/BadRequest'
          headers:
            Location:
              description: 'Redirect to the error page for the bad request.'
              type: string
              example: '/error/bad-request'
        '401':
          $ref: '#/components/responses/Unauthorized'
          headers:
            Location:
              description: 'Redirect to the login page or authorization error page.'
              type: string
              example: '/login'
        '404':
          $ref: '#/components/responses/NotFound'
          headers:
            Location:
              description: 'Redirect to the error page for the resource not found.'
              type: string
              example: '/error/not-found'
        '500':
          $ref: '#/components/responses/InternalServerError'
          headers:
            Location:
              description: 'Redirect to the general error page for server errors.'
              type: string
              example: '/error/internal-server'
  /events/search:
    get:
      operationId: R301
      summary: 'R301: Search Events API'
      description: 'Searches for events and returns the results as JSON. Access: MBR, ADM'

      tags: 
        - 'M03: Events and Tickets'

      parameters:
        - in: query
          name: search
          description: 'String to use for full-text search'
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
...
  /event/{event_id}/:
    get:
      operationId: R306
      summary: 'R306: View Event Details'
      description: 'Retrieves detailed information about a specific event, including media, ratings, and other necessary data. Access: MBR, ADM'
      tags:
        - 'M03: Events and Tickets'
      parameters:
        - in: path
          name: event_id
          schema:
            type: integer
          required: true
          description: 'Unique identifier of the event'
      responses:
        '200':
          description: 'Event details retrieved successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'               
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
 
  ## PURCHASE TICKET
  /event/{event_id}:
    post:
      operationId: R307
      summary: 'R307: Purchase Tickets for an Event'
      description: 'Allows a member to purchase a ticket for a specific event. Access: MBR, ART.'
      tags:
        - 'M03: Events and Tickets'
      parameters:
        - in: path
          name: event_id
          schema:
            type: integer
          required: true
          description: 'Unique identifier of the event'
      responses:
        '302':
          description: 'Redirect after processing the purchase information.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful authentication. Redirect to member ticket page.'
                  value: '/tickets'
                302Failure:
                  description: 'Failed authentication. Redirect to login form.'
                  value: '/login'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '409':
          description: 'Conflict. There are not enough tickets available or you have requested more tickets than the maximum allowed.'

  ## EDIT EVENT
...
components:
  schemas:
    Event:
      type: object
      properties:
        title:
          type: string
          description: 'Title of the event'
        description:
          type: string
          description: 'Detailed description of the event'
        type:
          type: string
          description: 'Visibility of the event (Public, Private)'
        date:
          type: string
          format: date-time
          description: 'Date and time of the event in ISO 8601 format'
        location:
          type: string
          description: 'Location of the event'
        media_url:
          type: string
          format: uri
          description: 'URL for event image'
        price:
          type: number
          format: float
          description: 'Price of the event'
        capacity:
          type: integer
          description: 'The maximum number of tickets available for this event'
        refund_policy:
          type: integer
          description: 'Refund percentage of the ticket price'
        artist_id:
          type: integer
          description: 'Owner of the event'
      example:
        title: 'Salsa Night with Live Music'
        description: 'Join us for an evening of vibrant salsa dancing and live music under the stars.'
        type: 'Public'
        date: '2024-12-15T20:00:00Z'
        location: 'Havana Club, Miami, USA'
        media_url: 'https://example.com/media/event101/image1.jpg'
        price: 20.00
        capacity: 500
        refund_policy: 100
...
    Ticket:
      type: object
      properties:
        ticket_id:
          type: integer
          description: 'Unique identifier of the ticket'
          required: true
        event_id:
          type: integer
          description: 'Unique identifier of the event'
          required: true
        purchase_date:
          type: string
          format: date-time
          description: 'Date and time the ticket was purchased'
          required: true
      example:
        ticket_id: 1001
        event_id: 301
        purchase_date: '2024-04-01T14:30:00Z'
  responses:
    BadRequest:
      description: 'Bad Request. The request could not be understood or was missing required parameters.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Invalid request parameters.'
    Unauthorized:
      description: 'Unauthorized. Authentication is required to access this resource.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'User is not authenticated.'
    NotFound:
      description: 'Not Found. The requested resource could not be found.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Resource not found.'
    InternalServerError:
      description: 'Internal Server Error. An unexpected error occurred on the server.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Something went wrong on the server.'